bpf/nat: implement support of ICMP6 Error packet too big
Running in masquerading BPF mode.

Assuming that, Pod is sending a packet to the world of a size bigger
than the accepted MTU in the networking path.

A networking equipment in that networking path would return an ICMP
Error packet with code PACKET_TOO_BIG. According to the RFE 5508,
that networking equipment should embed the original packet in its
response.

Previously because of the missing support of PACKET_TOO_BIG, the NAT
session to return the packet to Pod was never resolved. Meaning that
the ICMP Error packet got dropped.

 a) The src IP of the packet becomes the networking equipment that is
 generating the error response.  b) The protocol used for the response
 is ICMP where the packet sent may be TDP/UPD.

This change is updating the NAT process to read the embedded packet in
the ICMP Error packet that to determine the NAT session and send back
the packet to Pod.

This change is for IPv6.

Tests have been added in a way that the BPF_PROG will act as a
networking equipment sending the ICMP Error packet.

Fixes: #12968

Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@industrialdiscipline.com>